LayoutTests/inspector/styles/protocol-css-regions-commands-expected.txt

Tests the following commands and events:

getNamedFlowCollection Bug 91607
getFlowByName Bug 91855
namedFlowCreated Bug 92739
namedFlowRemoved Bug 92739


Running: testGetNamedFlowCollection
=== CSS Named Flows ===
flow1
flow2
flow3

Running: testGetFlowByName1
=== Named Flow "flow2" from main document ===
name: flow2
overset: false

Running: testGetFlowByName2
=== Name Flow "flow4" from main document ===
There is no Named Flow "flow4" in the main document

Running: testNamedFlowCreated
NamedFlowCreated: "tmpNamedFlow"

Running: testNamedFlowRemoved
NamedFlowRemoved: "tmpNamedFlow"
